+++ history.md
@@ -6,6 +6,7 @@
 - Enhancement to method `mpmToolbox.gui.MpmToolbox.closeProject()`, when the project is closed or a new project is loaded while playback is running, the playback will now stop properly.
 - Enhanced class `mpmToolbox.projectData.alignment.AbstractAlignmentComputation` and updated class `mpmToolbox.gui.audio.AudioDocumentData` accordingly. The latter has two additional buttons to select an alignment algorithm and execute it. 
   - Also class `mpmToolbox.projectData.alignment.PlaceholderAligner` has been added to illustrate the usage of the abstract class. This is basically a tutorial for anyone who wishes to integrate a new alignment method into MPM Toolbox.
+  - Enhancement in class `mpmToolbox.gui.audio.AudioDocumentData`. Some buttons on the bottom are disabled when they are not useful, e.g. when the piano roll displays a performance instead of an audio alignment or no audio is displayed.
 
 
 #### v0.1.21